index: add `ResolvedChangeId` to return visible and hidden commits
I'm going to add a way to select a specific commit from the set of
commits with a given change ID. We could sort these by committer
timestamp or something else, but index position is convenient because it
doesn't require reading the commits. The exact order isn't too
important, but having newer commits tend to have lower generation
numbers is nice because newer commits are used more frequently.