templates: show change generations for divergent and hidden commits
I replaced "??" with "divergent" because it gives a searchable term and
makes it clear that it's not a conflict, and I moved both "divergent"
and "hidden" to the end of the line in the log for consistency with
"conflict". This does make it less clear that it's related to the change
ID, but I think it makes it easier to parse visually since everything is
aligned better.